Keep BC in cache module
install PSR6 module
